% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_lite.R
\name{auc}
\alias{auc}
\title{Helper function to estimate AUC in the recovery plots}
\usage{
auc(rnk.list, max = NULL)
}
\arguments{
\item{rnk.list}{list of ranks for a particular annotation}

\item{max}{maximum rank}
}
\value{
AUC
}
\description{
Helper function to estimate AUC in the recovery plots
}
