% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrativeNMF-class.R
\docType{class}
\name{ButchR_integrativeNMF-class}
\alias{ButchR_integrativeNMF-class}
\alias{ButchR_integrativeNMF}
\title{Integrative NMF Class}
\value{
An object of class ButchR_integrativeNMF.
}
\description{
Integrative NMF Class
}
\section{Slots}{

\describe{
\item{\code{input_data}}{lsit of matrices}

\item{\code{HMatrix}}{list.}

\item{\code{HMatrix_vs}}{list.}

\item{\code{WMatrix_vs}}{list.}

\item{\code{FrobError}}{DataFrame.}

\item{\code{OptKStats}}{DataFrame.}

\item{\code{OptK}}{numeric}

\item{\code{SignFeatures}}{DataFrame or list}
}}

\examples{
\dontrun{
ButchR_integrativeNMF(input_data   = input_data,
                      HMatrix      = shared_HMatrix_list,
                      HMatrix_vs   = view_specific_HMatrix_list,
                      WMatrix_vs   = view_specific_WMatrix_list,
                      FrobError    = frob_errors,
                      OptKStats    = OptKStats,
                      OptK         = OptK,
                      SignFeatures = SignFeatures)
}
}
