% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinNMF-class.R
\docType{class}
\name{ButchR_joinNMF-class}
\alias{ButchR_joinNMF-class}
\alias{ButchR_joinNMF}
\title{Join NMF Class}
\value{
An object of class ButchR_joinNMF.
}
\description{
Join NMF Class
}
\section{Slots}{

\describe{
\item{\code{input_data}}{lsit of matrices}

\item{\code{HMatrix}}{list.}

\item{\code{WMatrix_vs}}{list.}

\item{\code{FrobError}}{DataFrame.}

\item{\code{OptKStats}}{DataFrame.}

\item{\code{OptK}}{numeric}

\item{\code{SignFeatures}}{DataFrame or list}
}}

\examples{
\dontrun{
ButchR_joinNMF(input_data   = input_data,
               HMatrix      = shared_HMatrix_list,
               WMatrix_vs   = view_specific_WMatrix_list,
               FrobError    = frob_errors,
               OptKStats    = OptKStats,
               OptK         = OptK,
               SignFeatures = SignFeatures)
}
}
