% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfExperiment-class_lite.R
\docType{class}
\name{ButchR_NMF-class}
\alias{ButchR_NMF-class}
\alias{ButchR_NMF}
\title{The ButchR_NMF class}
\value{
An object of class ButchR_NMF.
}
\description{
The ButchR_NMF class
}
\section{Slots}{

\describe{
\item{\code{input_matrix}}{list of matrices}

\item{\code{WMatrix}}{list.}

\item{\code{HMatrix}}{list.}

\item{\code{FrobError}}{DataFrame.}

\item{\code{OptKStats}}{DataFrame.}

\item{\code{OptK}}{numeric.}

\item{\code{SignFeatures}}{DataFrame or list.}
}}

\examples{
\dontrun{
ButchR_NMF(input_matrix = input_matrix,
           WMatrix      = lapply(complete_eval, "[[" , "W"),
           HMatrix      = lapply(complete_eval, "[[" , "H"),
           FrobError    = frob_errors,
           OptKStats    = OptKStats,
           OptK         = OptK,
           SignFeatures = SignFeatures)
}
}
