% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinNMF-class.R
\docType{methods}
\name{FrobError,join_NMF-method}
\alias{FrobError,join_NMF-method}
\title{Return Frobenius Error from all initializations}
\usage{
\S4method{FrobError}{join_NMF}(x, ...)
}
\arguments{
\item{x}{an nmfExperiment_lite object}
}
\value{
data.frame with Frobenius error for all factorization ranks
}
\description{
Return Frobenius Error from all initializations
}
\examples{
FrobError(jnmf_exp)
}
