% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/nmfExperiment-class_lite.R,
%   R/integrativeNMF-class.R, R/joinNMF-class.R
\docType{methods}
\name{FrobError}
\alias{FrobError}
\alias{FrobError,nmfExperiment_lite-method}
\alias{FrobError,ANY-method}
\alias{FrobError,integrative_NMF-method}
\alias{FrobError,join_NMF-method}
\title{Frobenius error for all factorization ranks}
\usage{
FrobError(x, ...)

\S4method{FrobError}{nmfExperiment_lite}(x, ...)

\S4method{FrobError}{integrative_NMF}(x, ...)

\S4method{FrobError}{join_NMF}(x, ...)
}
\arguments{
\item{x}{an R object of type ButchR_NMF, ButchR_JoinNMF, or
ButchR_IntegrativeNMF.}

\item{...}{additional parameters.}
}
\value{
data.frame with Frobenius errors
}
\description{
Returns a data.frame with factorization ranks in the columns,
and the Frobenius error for every NMF initialization in the rows.
}
\examples{
data("leukemia")
nmf_exp <- runNMFtensor_lite(leukemia$matrix, ranks = 2,
method = "NMF",
n_initializations = 2)
FrobError(nmf_exp)
\dontrun{
data("leukemia")
nmf_exp <- runNMFtensor_lite(leukemia$matrix, ranks = 2:10,
method = "NMF",
n_initializations = 2)
FrobError(nmf_exp)
}
}
