% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfExperiment-class_lite.R
\docType{methods}
\name{FrobError,nmfExperiment_lite-method}
\alias{FrobError,nmfExperiment_lite-method}
\title{Return Frobenius Error from all initializations}
\usage{
\S4method{FrobError}{nmfExperiment_lite}(x, ...)
}
\arguments{
\item{x}{an nmfExperiment_lite object}
}
\value{
data.frame with Frobenius error for all factorization ranks
}
\description{
Return Frobenius Error from all initializations
}
\examples{
FrobError(nmf_exp)
}
