% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/plot_functions_lite.R
\docType{methods}
\name{generateRiverplot}
\alias{generateRiverplot}
\alias{generateRiverplot,ButchR_NMF-method}
\alias{generateRiverplot,ANY,ANY-method}
\title{NMF riverplot}
\usage{
generateRiverplot(nmf_exp, ...)

\S4method{generateRiverplot}{ButchR_NMF}(
  nmf_exp,
  edges.cutoff = 0,
  useH = FALSE,
  color = TRUE,
  ranks = NULL
)
}
\arguments{
\item{nmf_exp}{an object of class ButchR_NMF.}

\item{...}{additional parameters.}

\item{edges.cutoff}{cutoff until which similarities are displayed}

\item{useH}{whether to relate signatures (FALSE) or exposures (TRUE)}

\item{color}{whether to colorize the nodes based on PCA of the signatures}

\item{ranks}{numeric vector of the selected factorization ranks to generate
a riverplot.}
}
\value{
ret riverplot object
}
\description{
Generate a riverplot object that displays the similarities between
signatures at different factorization ranks
}
\examples{
\dontrun{
plt <- generateRiverplot(nmf_exp, edges.cutoff = 0.2)
plot(plt, plot_area = 1, yscale = 0.6, nodewidth = 0.5)
}
\dontrun{
data("leukemia")
nmf_exp <- run_NMF_tensor(leukemia$matrix, ranks = 2:10,
method = "NMF",
n_initializations = 2)
plot(generateRiverplot(nmf_exp))
plot(generateRiverplot(nmf_exp, ranks = 2:5))
}
}
