% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_lite.R
\name{gg_plotKStats}
\alias{gg_plotKStats}
\title{Plots optimal K metrics}
\usage{
gg_plotKStats(
  nmf_exp,
  plot_vars = c("FrobError", "FrobError_cv", "sumSilWidth", "meanSilWidth",
    "copheneticCoeff", "meanAmariDist")
)
}
\arguments{
\item{nmf_exp}{an object of class ButchR_NMF, ButchR_joinNMF, or
ButchR_integrativeNMF.}

\item{plot_vars}{a character vector with the ids of the metrics to display.
possible values are: FrobError, FrobError_min, FrobError_mean, FrobError_cv,
FrobError_sd, sumSilWidth, meanSilWidth, copheneticCoeff, and meanAmariDist.
Default value: c("FrobError", "FrobError_cv", "sumSilWidth", "meanSilWidth",
"copheneticCoeff", "meanAmariDist").}
}
\value{
a ggplot figure with the values for the selected factorization metrics
}
\description{
For every factorization rank the Frobenius error,
coefficient variation of Frobenius error,
sum Silhouette Width, mean Silhouette width,
cophenetic coefficient and mean Amari distance is shown.
}
\examples{
\dontrun{
data("leukemia")
nmf_exp <- run_NMF_tensor(leukemia$matrix, ranks = 2:10,
method = "NMF",
n_initializations = 2)
gg_plotKStats(nmf_exp)
}
}
