% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinNMF.R
\docType{methods}
\name{HMatrix,join_NMF-method}
\alias{HMatrix,join_NMF-method}
\title{Return a list of H-Matrices or an H-Matrix for the indicaded rank}
\usage{
\S4method{HMatrix}{join_NMF}(x, k = NULL, ...)
}
\arguments{
\item{x}{a join_NMF object}

\item{k}{numeric - factorization rank}
}
\value{
list of H-Matrices or an H-Matrix for the indicaded rank
}
\description{
Return a list of H-Matrices or an H-Matrix for the indicaded rank
}
\examples{
HMatrix(jnmf_exp)
HMatrix(jnmf_exp, k = 2)
}
