% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfExperiment-class_lite.R
\docType{methods}
\name{HMatrix,nmfExperiment_lite-method}
\alias{HMatrix,nmfExperiment_lite-method}
\title{Return a list of H-Matrices or an H-Matrix for the indicaded rank}
\usage{
\S4method{HMatrix}{nmfExperiment_lite}(x, k = NULL, ...)
}
\arguments{
\item{x}{an nmfExperiment_lite object}

\item{k}{numeric  - factorization rank}
}
\value{
list of H-Matrices or an H-Matrix for the indicaded rank
}
\description{
Return a list of H-Matrices or an H-Matrix for the indicaded rank
}
\examples{
HMatrix(nmf_exp)
HMatrix(nmf_exp, k = 2)
}
