% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_functions.R
\docType{class}
\name{integrative_NMF-class}
\alias{integrative_NMF-class}
\alias{integrative_NMF}
\title{Integrative NMF Class}
\value{

}
\description{
Integrative NMF Class
}
\section{Slots}{

\describe{
\item{\code{shared_HMatrix_list}}{list.}

\item{\code{view_specific_HMatrix_list}}{list.}

\item{\code{view_specific_WMatrix_list}}{list.}

\item{\code{best_factorization_idx}}{numeric}

\item{\code{view_specific_NMFexp_list}}{list}
}}

