% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinNMF-class.R
\docType{class}
\name{join_NMF-class}
\alias{join_NMF-class}
\alias{join_NMF}
\title{Join NMF Class}
\value{

}
\description{
Join NMF Class
}
\section{Slots}{

\describe{
\item{\code{HMatrix}}{list.}

\item{\code{WMatrix_vs}}{list.}

\item{\code{FrobError}}{DataFrame.}

\item{\code{OptKStats}}{DataFrame.}
}}

\examples{
\dontrun{
join_NMF(input_data   = input_data,
         HMatrix      = shared_HMatrix_list,
         WMatrix_vs   = view_specific_WMatrix_list,
         FrobError    = frob_errors,
         OptKStats    = OptKStats,
         OptK         = OptK,
         SignFeatures = SignFeatures)
}
}
