% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{leukemia}
\alias{leukemia}
\title{Microarray expression data of B-ALL, T-ALL and AML samples.}
\format{
A list with an expression matrix and annotatin data frame
\describe{
  \item{matrix}{4452 rows/genes and 38 samples}
  \item{annotation}{AL and _AML status with type}
  ...
}
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/10521349/}
}
\usage{
data(leukemia)
}
\description{
Dataset containing Affymetrix Hgu6800 microarray expression data of B-ALL,
T-ALL and AML samples.
This dataset had also been used by Brunet et al. (PNAS, 2004) and
Gaujoux et al. (BMC Bioinformatics, 2010).
}
\keyword{datasets}
