% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfExperiment-class_lite.R
\name{normalize_matrix_per_dim}
\alias{normalize_matrix_per_dim}
\title{Normalize matrix by columns}
\usage{
normalize_matrix_per_dim(in_matrix, in_dimension)
}
\arguments{
\item{in_matrix}{matrix to normalize.}

\item{in_dimension}{normalize by rows == 1 or columns == 1 .}
}
\value{
matrix normalized by columns, colsum == 1.
}
\description{
Normalize matrix by columns
}
\examples{
\dontrun{
data("leukemia")
nmf_exp <- runNMFtensor_lite(leukemia$matrix, ranks = 2:10,
method = "NMF",
n_initializations = 10)
normalize_matrix_per_dim(WMatrix(nmf_exp, k=2), 2)
}
}
