% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/nmfExperiment-class_lite.R
\docType{methods}
\name{normalizeH}
\alias{normalizeH}
\alias{normalizeH,ButchR_NMF-method}
\alias{normalizeH,ANY,ANY-method}
\title{Normalize the signatures matrix (H)}
\usage{
normalizeH(nmf_exp, ...)

\S4method{normalizeH}{ButchR_NMF}(nmf_exp)
}
\arguments{
\item{nmf_exp}{an object of type ButchR_NMF.}

\item{...}{additional parameters.}
}
\value{
a ButchR_NMF object normalized by H.
}
\description{
After row normalization of the matrix H, the inverse factors are
mutiplied with the columns of W in order to keep the matrix product W*H
constant.
}
\examples{
data("leukemia")
nmf_exp <- runNMFtensor_lite(leukemia$matrix, ranks = 2,
method = "NMF",
n_initializations = 2)
normalizeH(nmf_exp)
\dontrun{
data("leukemia")
nmf_exp <- runNMFtensor_lite(leukemia$matrix, ranks = 2:10,
method = "NMF",
n_initializations = 10)
normalizeH(nmf_exp)
}
}
