% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfExperiment-class_lite.R
\name{OptK,ButchR_NMF-method}
\alias{OptK,ButchR_NMF-method}
\title{Return optimal K}
\usage{
\S4method{OptK}{ButchR_NMF}(x, ...)
}
\arguments{
\item{x}{an object of class ButchR_NMF.}

\item{...}{additional parameters.}
}
\value{
numeric - optimal K
}
\description{
Return optimal K
}
\examples{
\dontrun{
data("leukemia")
nmf_exp <- runNMFtensor_lite(leukemia$matrix, ranks = 2:10,
method = "NMF",
n_initializations = 10)
OptK(nmf_exp)
}
}
