% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfExperiment-class.R,
%   R/nmfExperiment-class_lite.R
\docType{methods}
\name{OptK,nmfExperiment-method}
\alias{OptK,nmfExperiment-method}
\title{Optimal K}
\usage{
\S4method{OptK}{nmfExperiment}(x, ...)

\S4method{OptK}{nmfExperiment}(x, ...)
}
\arguments{
\item{x}{an nmfExperiment_lite object}

\item{nmfExperiment}{}
}
\value{
numeric - optimal K
}
\description{
Optimal K

Return optimal K
}
\examples{
OptK(nmf_exp)
}
