% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfExperiment-class_lite.R
\name{OptK,nmfExperiment_lite-method}
\alias{OptK,nmfExperiment_lite-method}
\title{Return optimal K}
\usage{
\S4method{OptK}{nmfExperiment_lite}(x, ...)
}
\arguments{
\item{x}{an nmfExperiment_lite object}

\item{...}{additional parameters.}
}
\value{
numeric - optimal K
}
\description{
Return optimal K
}
\examples{
\dontrun{
data("leukemia")
nmf_exp <- runNMFtensor_lite(leukemia$matrix, ranks = 2:10,
method = "NMF",
n_initializations = 10)
OptK(nmf_exp)
}
}
