% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfExperiment-class.R,
%   R/nmfExperiment-class_lite.R
\docType{methods}
\name{OptKStats,nmfExperiment-method}
\alias{OptKStats,nmfExperiment-method}
\title{Optimal K Statistics getter}
\usage{
\S4method{OptKStats}{nmfExperiment}(x, ...)

\S4method{OptKStats}{nmfExperiment}(x, ...)
}
\arguments{
\item{x}{an nmfExperiment_lite object}

\item{nmfExperiment}{}
}
\value{
optimal K Statistics
}
\description{
Optimal K Statistics getter

Return optimal K Statistics
}
\examples{
OptKStats(nmf_exp)
}
