% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfExperiment-class_lite.R
\name{OptKStats,nmfExperiment_lite-method}
\alias{OptKStats,nmfExperiment_lite-method}
\title{Return optimal factorization rank (K) Statistics}
\usage{
\S4method{OptKStats}{nmfExperiment_lite}(x, ...)
}
\arguments{
\item{x}{an nmfExperiment_lite object}
}
\value{
optimal K Statistics
}
\description{
Return optimal factorization rank (K) Statistics
}
\examples{
\dontrun{
data("leukemia")
nmf_exp <- runNMFtensor_lite(leukemia$matrix, ranks = 2:10,
method = "NMF",
n_initializations = 2)
OptKStats(nmf_exp)
}
}
