% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_functions.R
\name{rankTransform}
\alias{rankTransform}
\title{Function to rankTransforma matrix}
\usage{
rankTransform(matrix)
}
\arguments{
\item{matrix}{input matrix}
}
\value{

}
\description{
Function to rankTransforma matrix
}
\examples{
\dontrun{
h <- HMatrix(nmf_exp, k = 5)
rankTransform(h)
}
}
