% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{recovery_plot}
\alias{recovery_plot}
\title{Recovery plot}
\usage{
recovery_plot(x, annot, ...)
}
\arguments{
\item{x}{an object of class ButchR_NMF, ButchR_joinNMF, or
ButchR_integrativeNMF.}

\item{annot}{input named factor/character with biological/clinical
annotation for the columns of the input data.}

\item{...}{additional parameters.}
}
\value{
recovery plot or a data frame with the association analysis results.
}
\description{
Generate a recovery plot to visualize the association between the NMF
signatures and a biological or clinical annotation factor.
}
\examples{
\dontrun{
recovery_plot(x, annot_factor)
}
}
