% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_lite.R
\name{relabelRiverplot}
\alias{relabelRiverplot}
\title{Relabel and recolour a riverplot}
\usage{
relabelRiverplot(in_riverplot, in_list)
}
\arguments{
\item{in_riverplot}{riverplot object}

\item{in_list}{list of new colors and labels to use}
}
\value{
relabeled riverplots
}
\description{
Relabel and recolour a riverplot
}
