% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_lite.R
\name{reorderEdges}
\alias{reorderEdges}
\title{Order the riverplot edges to prevent crossings from a single node}
\usage{
reorderEdges(nodes, edges)
}
\arguments{
\item{nodes}{riverplot nodes}

\item{edges}{riverplot nodes#'}
}
\value{
edges
}
\description{
Order the riverplot edges to prevent crossings from a single node
}
