% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_functions.R
\name{run_integrative_NMF_tensor}
\alias{run_integrative_NMF_tensor}
\title{Computes integrative NMF on tensorflow using the reticulate framework}
\usage{
run_integrative_NMF_tensor(matrix_list, k_min = 2, k_max = 2,
  outer_iter = 10, inner_iter = 10^4, conver_stop_threshold = 40,
  lambda = 1, Sp = 0)
}
\arguments{
\item{lambda}{}
}
\value{

}
\description{
Computes integrative NMF on tensorflow using the reticulate framework
}
