% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinNMF_functions.R
\name{run_joinNMF_tensor}
\alias{run_joinNMF_tensor}
\title{Join NMF}
\usage{
run_joinNMF_tensor(
  matrix_list,
  ranks = 2,
  n_initializations = 10,
  iterations = 10^4,
  convergence_threshold = 40,
  Sp = 0,
  extract_features = FALSE
)
}
\arguments{
\item{matrix_list}{List of non-negative matrices}

\item{ranks}{numeric vector with ranks to factorize}

\item{n_initializations}{Number of initializations to evaluate}

\item{iterations}{Maximum number of iterations to run for every initialization}

\item{convergence_threshold}{The factorization stops,
if the convergence test is constant for this number of iterations}

\item{Sp}{Sparsity,}

\item{extract_features}{if TRUE performs feature extraction for all
factorization ranks > 2.}
}
\value{
An object of class ButchR_joinNMF.
containing a join H matrix and one W matrix for each input matrix
}
\description{
Computes join NMF on tensorflow using the reticulate framework,
uses a list of non-negative matrices as input
}
\examples{
\dontrun{
norm_mat_list <- list(a = matrix(abs(rnorm(1000)), ncol = 10),
                      b = matrix(abs(rnorm(1000)), ncol = 10))
jnmf_exp <- run_joinNMF_tensor(norm_mat_list,
                               ranks = 2:5,
                               n_initializations     = 10,
                               iterations            = 10^4,
                               convergence_threshold = 40)
jnmf_exp
}
}
