% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_functions_lite.R
\name{run_NMF_tensor}
\alias{run_NMF_tensor}
\title{Single view NMF}
\usage{
run_NMF_tensor(
  X,
  ranks,
  method = "NMF",
  n_initializations = 10,
  iterations = 10^4,
  convergence_threshold = 40,
  n_neighbors = 4,
  alpha = 0.1,
  lamb = 10,
  graph = NULL,
  seed = NULL,
  extract_features = FALSE
)
}
\arguments{
\item{X}{Input matrix, should be a non-negative matrix}

\item{ranks}{numeric vector with ranks to factorize}

\item{method}{method to use in the factorization, available:
"NMF", "GRNMF_SC"}

\item{n_initializations}{Number of initializations to evaluate}

\item{iterations}{Maximum number of iterations to run for every initialization}

\item{convergence_threshold}{The factorization stops, if the convergence test
constant for this number of iterations}

\item{n_neighbors}{for method "GRNMF_SC", the number of neighbors to take
into account when building the graph G}

\item{alpha}{for method "GRNMF_SC", regularization parameter alpha}

\item{lamb}{for method "GRNMF_SC", regularization parameter lambda}

\item{graph}{for method "GRNMF_SC", square matrix representing a graph
between columns of the input matrix, values correspond to edge weight,
if NULL will compute own graph.}

\item{seed}{Numeric seed to initialize matrices W and H.}

\item{extract_features}{if TRUE performs feature extraction for all f
actorization ranks > 2.}
}
\value{
A ButchR_NMF object, containg the factorized matrices W and
 H, along with factorization metrics
}
\description{
Computes NMF on tensorflow using the reticulate framework, uses a
non-negative matrix as input
}
\examples{
nmf_exp <- run_NMF_tensor(matrix(1:1000, ncol = 10),
                             ranks = 2:5,
                             n_initializations     = 10,
                             iterations            = 10^4,
                             convergence_threshold = 40)
nmf_exp
}
