% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_functions.R
\name{runNMFtensor}
\alias{runNMFtensor}
\title{Computes NMF on tensorflow using the reticulate framework}
\usage{
runNMFtensor(nmf.exp, k.min = 2, k.max = 2, outer.iter = 10,
  inner.iter = 10^4, conver.test.niter = 10,
  conver.test.stop.threshold = 40)
}
\arguments{
\item{conver.test.stop.threshold}{}
}
\value{

}
\description{
Computes NMF on tensorflow using the reticulate framework
}
