% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_functions_lite.R
\name{runNMFtensor_lite}
\alias{runNMFtensor_lite}
\title{Single view NMF}
\usage{
runNMFtensor_lite(X, ranks, n_initializations = 10, iterations = 10^4,
  convergence_threshold = 40)
}
\arguments{
\item{X}{Input matrix, should be a non-negative matrix}

\item{n_initializations}{Number of initializations to evaluate}

\item{iterations}{Maximum number of iterations to run for every initialization}

\item{convergence_threshold}{The factorization stops, if the convergence test is constant for this number of iterations}
}
\value{
A nmfExperiment_lite object, containg the initial matrix X and the faactorized matrices W and H, along with factorization metrics
}
\description{
Computes NMF on tensorflow using the reticulate framework, uses a non-negative matrix as input
}
\examples{

}
