% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_functions.R
\name{sigmoidTransform}
\alias{sigmoidTransform}
\title{Function to used to perform a k-means k=2 on a W matrix}
\usage{
sigmoidTransform(col.vector, q = 0.95)
}
\arguments{
\item{col.vector}{vector taken from a matrix columns}

\item{q}{quantile}
}
\value{
vector
}
\description{
Function to used to perform a k-means k=2 on a W matrix
}
\examples{
\dontrun{
Wf <- HMatrix(nmf_exp, k = 5)
ssf <- apply(Wf, 1, function(x){
x <- sigmoidTransform(x)
k <- kmeans(x, 2)
max_idx <- which.max(k$centers)
paste(if_else(k$cluster == max_idx, "1", "0"), collapse = "")
})
}
}
