% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/nmfExperiment-class_lite.R
\docType{methods}
\name{SignatureFeatures}
\alias{SignatureFeatures}
\alias{SignatureFeatures,ButchR_NMF-method}
\alias{SignatureFeatures,ANY-method}
\title{Signature Features}
\usage{
SignatureFeatures(x, k = NULL, ...)

\S4method{SignatureFeatures}{ButchR_NMF}(x, k = NULL, ...)
}
\arguments{
\item{x}{an object of class ButchR_NMF, ButchR_joinNMF, or
ButchR_integrativeNMF.}

\item{k}{numeric  - factorization rank}

\item{...}{additional parameters.}
}
\value{
list of binary matrices if `k`=NULL or a single binary matrix for
the selected factorization rank.
}
\description{
Returns a binary matrix for every factorization rank,
with features in the rows and samples in the columns,
in which 1 means that the features is contributing to the signature,
and 0 it does not.
The extraction of Signature Features is not supported for k = 2.
}
\examples{
data("leukemia")
nmf_exp <- run_NMF_tensor(leukemia$matrix, ranks = 3,
                             method = "NMF",
                             n_initializations = 2,
                             extract_features = TRUE)
SignatureFeatures(nmf_exp)
SignatureFeatures(nmf_exp, k = 3)
\dontrun{
# For ButchR_NMF objects:
SignatureFeatures(nmf_exp)
SignatureFeatures(nmf_exp, k = 3)
}
}
