% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{SignatureSpecificFeatures}
\alias{SignatureSpecificFeatures}
\title{Signature Specific Features}
\usage{
SignatureSpecificFeatures(x, k = NULL, return_all_features = FALSE,
  ...)
}
\arguments{
\item{x}{an nmfExperiment or a nmfExperiment_lite object}

\item{k}{numeric  - factorization rank}
}
\value{
list of signature specific fatures or binary matrices for all features
}
\description{
Returns the list of signatures specific features
for all factorization ranks or for the indicaded rank,
if return_all_features = TRUE
returns a binary matrix for every factorization rank,
with features in the rows and samples in the columns,
in which 1 means that the features is contributing to the signature,
and 0 it does not.
The extraction of Signature Specific Features is not supported for k = 2
}
\examples{
SignatureSpecificFeatures(nmf_exp)
SignatureSpecificFeatures(nmf_exp, k = 2)
SignatureSpecificFeatures(nmf_exp, k = 2, return_all_features = TRUE)
}
