% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfExperiment-class_lite.R
\docType{methods}
\name{SignatureSpecificFeatures,nmfExperiment_lite-method}
\alias{SignatureSpecificFeatures,nmfExperiment_lite-method}
\title{Signature specfific features getter}
\usage{
\S4method{SignatureSpecificFeatures}{nmfExperiment_lite}(x, k = NULL,
  return_all_features = FALSE, ...)
}
\arguments{
\item{nmfExperiment_lite}{}
}
\description{
Signature specfific features getter
}
\examples{
SignatureSpecificFeatures(nmf_exp)
lapply(SignatureSpecificFeatures(nmf_exp), function(x) sapply(x, length))
SignatureSpecificFeatures(nmf_exp, k = 3)
SignatureSpecificFeatures(nmf_exp, k = 3, return_all_features = TRUE)
}
