% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_functions_lite.R
\name{source_NMFtensor_function}
\alias{source_NMFtensor_function}
\title{Single view NMF implementations}
\usage{
source_NMFtensor_function(method)
}
\arguments{
\item{method}{string with method to use, "NMF", "GRNMF_SC"}
}
\value{
python function to carry out the factorization on tensorflow
}
\description{
Python functions to run several NMF implementations on tensorflow
using the reticulate framework, uses a non-negative matrix as input
}
\examples{
x <- source_NMFtensor_function("NMF")
x
# Bratwurst:::source_NMFtensor_function("NMF")
}
