% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation.R
\name{val_graph_GRNMF_SC}
\alias{val_graph_GRNMF_SC}
\title{Validate if input graph is a matrix for GRNMF_SC}
\usage{
val_graph_GRNMF_SC(graph, X, method)
}
\arguments{
\item{graph}{graph}

\item{X}{matrix}

\item{method}{NMF method}
}
\value{
graph
}
\description{
Validate if input graph is a matrix for GRNMF_SC
}
\examples{
\dontrun{
ButchR:::val_graph_GRNMF_SC(graph)
}
}
