% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation.R
\name{val_nonnegative_matrix}
\alias{val_nonnegative_matrix}
\title{Validate if input matrix for NMF}
\usage{
val_nonnegative_matrix(X)
}
\arguments{
\item{X}{matrix}
}
\value{
X
}
\description{
Validate if input matrix for NMF
}
\examples{
\dontrun{
ButchR:::val_nonnegative_matrix(X)
}
}
