% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation.R
\name{val_positive_numeric}
\alias{val_positive_numeric}
\title{Validate input positive numeric vector}
\usage{
val_positive_numeric(x, id)
}
\arguments{
\item{x}{numeric}

\item{id}{param id}
}
\description{
Validate input positive numeric vector
}
\examples{
\dontrun{
ButchR:::val_positive_numeric(X)
}
}
