% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation.R
\name{val_ranks_torun}
\alias{val_ranks_torun}
\title{Validate input ranks}
\usage{
val_ranks_torun(ranks, ncolmat)
}
\arguments{
\item{ranks}{numeric with ranks}

\item{ncolmat}{ncol of input matrix}
}
\description{
Validate input ranks
}
\examples{
\dontrun{
ButchR:::val_ranks_torun(X)
}
}
