% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation.R
\name{val_single_boolean}
\alias{val_single_boolean}
\title{Validate input single boolean}
\usage{
val_single_boolean(x, id)
}
\arguments{
\item{x}{bool}

\item{id}{param id}
}
\description{
Validate input single boolean
}
\examples{
\dontrun{
ButchR:::val_single_numeric(X)
}
}
