% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation.R
\name{val_single_integer}
\alias{val_single_integer}
\title{Validate input single numeric integer}
\usage{
val_single_integer(x, id, minval = 0)
}
\arguments{
\item{x}{numeric}

\item{id}{param id}

\item{minval}{minimum value to test}
}
\description{
Validate input single numeric integer
}
\examples{
\dontrun{
ButchR:::val_single_integer(X)
}
}
