% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation.R
\name{val_single_numeric}
\alias{val_single_numeric}
\title{Validate input single numeric}
\usage{
val_single_numeric(x, id)
}
\arguments{
\item{x}{numeric}

\item{id}{param id}
}
\description{
Validate input single numeric
}
\examples{
\dontrun{
ButchR:::val_single_numeric(X)
}
}
