% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_functions_lite.R
\name{WcomputeFeatureStats}
\alias{WcomputeFeatureStats}
\title{Computes Signature specific features}
\usage{
WcomputeFeatureStats(W)
}
\arguments{
\item{W}{W matrix with more than 2 signatures}
}
\description{
Computes Signature specific features
}
\examples{
\dontrun{
WcomputeFeatureStats(W)
}
}
