% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinNMF.R
\docType{methods}
\name{WMatrix,join_NMF-method}
\alias{WMatrix,join_NMF-method}
\title{Return a list of W-Matrices or a W-Matrix for the indicaded rank}
\usage{
\S4method{WMatrix}{join_NMF}(x, k = NULL, ...)
}
\arguments{
\item{x}{an join_NMF object}

\item{k}{numeric  - factorization rank}
}
\value{
list of W-Matrices or a W-Matrix for the indicaded rank
}
\description{
Return a list of W-Matrices or a W-Matrix for the indicaded rank
}
\examples{
WMatrix(jnmf_exp)
WMatrix(jnmf_exp, k = 2)
}
