% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/nmfExperiment-class_lite.R,
%   R/integrativeNMF-class.R, R/joinNMF-class.R
\docType{methods}
\name{WMatrix}
\alias{WMatrix}
\alias{WMatrix,ButchR_NMF-method}
\alias{WMatrix,ANY-method}
\alias{WMatrix,integrative_NMF-method}
\alias{WMatrix,ButchR_joinNMF-method}
\title{W-Matrix (W-Matrix with smallest frobError)}
\usage{
WMatrix(x, k = NULL, ...)

\S4method{WMatrix}{ButchR_NMF}(x, k = NULL, ...)

\S4method{WMatrix}{integrative_NMF}(x, k = NULL, view_id = NULL, ...)

\S4method{WMatrix}{ButchR_joinNMF}(x, k = NULL, view_id = NULL, ...)
}
\arguments{
\item{x}{an object of type ButchR_NMF, ButchR_joinNMF, or
ButchR_integrativeNMF.}

\item{k}{numeric - factorization rank.}

\item{...}{additional parameters.}

\item{view_id}{character vector with views from which to extract W matrices}
}
\value{
list of W-Matrices or a W-Matrix for the indicated rank.
}
\description{
Return a list of W-Matrices or a W-Matrix for the indicaded rank
}
\examples{
data("leukemia")
nmf_exp <- runNMFtensor_lite(leukemia$matrix, ranks = 2,
method = "NMF",
n_initializations = 2)
WMatrix(nmf_exp)
WMatrix(nmf_exp, k = 2)
# For ButchR_NMF objects:
WMatrix(nmf_exp)
WMatrix(nmf_exp, k = 2)
\dontrun{
# For integrative_NMF objects:
WMatrix(inmf_exp)
WMatrix(inmf_exp, k = 2)
lapply(WMatrix(inmf_exp, k = 2), head)
WMatrix(inmf_exp, k = 2, view_id = "atac")
}
\dontrun{
# For ButchR_joinNMF objects:
WMatrix(jnmf_exp)
WMatrix(jnmf_exp, k = 2)
lapply(WMatrix(jnmf_exp, k = 2), head)
WMatrix(jnmf_exp, k = 2, view_id = "atac")
}
}
