% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfExperiment-class_lite.R
\docType{methods}
\name{WMatrix,nmfExperiment_lite-method}
\alias{WMatrix,nmfExperiment_lite-method}
\title{Return a list of W-Matrices or a W-Matrix for the indicaded rank}
\usage{
\S4method{WMatrix}{nmfExperiment_lite}(x, k = NULL, ...)
}
\arguments{
\item{x}{an nmfExperiment_lite object}

\item{k}{numeric  - factorization rank}
}
\value{
list of W-Matrices or a W-Matrix for the indicaded rank
}
\description{
Return a list of W-Matrices or a W-Matrix for the indicaded rank
}
\examples{
WMatrix(nmf_exp)
WMatrix(nmf_exp, k = 2)
}
