% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_functions_lite.R
\name{compute_OptKStats_NMF}
\alias{compute_OptKStats_NMF}
\title{Computes factorization optimal K stats}
\usage{
compute_OptKStats_NMF(complete_eval)
}
\arguments{
\item{complete_eval}{internal object return after computing NMF with tensorflow}
}
\description{
Computes factorization optimal K stats
}
