% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/integrativeNMF-class.R
\docType{methods}
\name{compute_SignatureFeatures}
\alias{compute_SignatureFeatures}
\alias{compute_SignatureFeatures,integrative_NMF-method}
\alias{compute_SignatureFeatures,ANY-method}
\title{Compute Signature Features
Performs feature extraction for all
factorization ranks > 2.
Returns an nmfExperiment object with computed signature features}
\usage{
compute_SignatureFeatures(x)

\S4method{compute_SignatureFeatures}{integrative_NMF}(x)
}
\arguments{
\item{x}{an nmfExperiment, integrative_NMF objects}
}
\value{
the same type of input object with computed signature features.
}
\description{
A K-means with k=2 is run over all rows of the W matrices.
Resulting in a binary matrix for every factorization rank,
with features in the rows and samples in the columns,
in which 1 means that the features is contributing to the signature,
and 0 it does not.
The extraction of Signature Specific Features is not supported for k = 2
}
\examples{
\dontrun{
data("leukemia")
nmf_exp <- runNMFtensor_lite(leukemia$matrix, ranks = 3,
                             method = "NMF",
                             n_initializations = 2,
                             extract_features = FALSE)
nmf_exp <- compute_SignatureFeatures(nmf_exp)
SignatureSpecificFeatures(nmf_exp, k = 3)
SignatureSpecificFeatures(nmf_exp, k = 3, return_all_features = TRUE)
}
\dontrun{
data("leukemia")
nmf_exp <- runNMFtensor_lite(leukemia$matrix, ranks = 3,
                             method = "NMF",
                             n_initializations = 2,
                             extract_features = FALSE)
nmf_exp <- compute_SignatureFeatures(nmf_exp)
SignatureSpecificFeatures(nmf_exp, k = 3)
SignatureSpecificFeatures(nmf_exp, k = 3, return_all_features = TRUE)
}
}
