% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_functions_lite.R
\name{cosineDissMat}
\alias{cosineDissMat}
\title{Create distance matrix with cosine similarity with matrix operations}
\usage{
cosineDissMat(in.matrix, in.dimension = 2)
}
\arguments{
\item{in.matrix}{concat_matrix a concatenated W matrix across multiple initializations}

\item{in.dimension}{factorization rank}
}
\description{
Create distance matrix with cosine similarity with matrix operations
}
\examples{
\dontrun{
# concat_matrix a concatenated W matrix across multiple initializations
cosineDissMat(as.matrix(concat_matrix))
}
}
