% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/nmfExperiment-class_lite.R,
%   R/integrativeNMF-class.R, R/joinNMF-class.R
\docType{methods}
\name{FrobError}
\alias{FrobError}
\alias{FrobError,ButchR_NMF-method}
\alias{FrobError,ANY-method}
\alias{FrobError,ButchR_integrativeNMF-method}
\alias{FrobError,ButchR_joinNMF-method}
\title{Frobenius error for all factorization ranks}
\usage{
FrobError(x, ...)

\S4method{FrobError}{ButchR_NMF}(x, ...)

\S4method{FrobError}{ButchR_integrativeNMF}(x, ...)

\S4method{FrobError}{ButchR_joinNMF}(x, ...)
}
\arguments{
\item{x}{an object of type ButchR_NMF, ButchR_joinNMF, or
ButchR_integrativeNMF.}

\item{...}{additional parameters.}
}
\value{
data.frame with Frobenius errors
}
\description{
Returns a data.frame with factorization ranks in the columns,
and the Frobenius error for every NMF initialization in the rows.
}
\examples{
data("leukemia")
nmf_exp <- run_NMF_tensor(leukemia$matrix, ranks = 2,
method = "NMF",
n_initializations = 2)
FrobError(nmf_exp)
\dontrun{
data("leukemia")
nmf_exp <- run_NMF_tensor(leukemia$matrix, ranks = 2:10,
method = "NMF",
n_initializations = 2)
FrobError(nmf_exp)
}
}
