% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_lite.R
\name{gg_plotKStats}
\alias{gg_plotKStats}
\title{Plots optimal K metrics}
\usage{
gg_plotKStats(
  nmf_exp,
  plot_vars = c("FrobError", "cv", "sumSilWidth", "meanSilWidth", "copheneticCoeff",
    "meanAmariDist")
)
}
\arguments{
\item{nmf_exp}{an object of class ButchR_NMF.}

\item{plot_vars}{character - ids of the metrics to display}
}
\value{
a ggplot figure with the values for the six factorization metrics
}
\description{
For every factorization rank the Frobenius error,
coefficient variation of Frobenius error,
sum Silhouette Width, mean Silhouette width,
cophenetic coefficient and mean Amari distance is shown
}
\examples{
\dontrun{
data("leukemia")
nmf_exp <- runNMFtensor_lite(leukemia$matrix, ranks = 2:10,
method = "NMF",
n_initializations = 2)
gg_plotKStats(nmf_exp)
}
}
