% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/nmfExperiment-class_lite.R,
%   R/integrativeNMF-class.R, R/joinNMF-class.R
\docType{methods}
\name{HMatrix}
\alias{HMatrix}
\alias{HMatrix,ButchR_NMF-method}
\alias{HMatrix,ANY-method}
\alias{HMatrix,ButchR_integrativeNMF-method}
\alias{HMatrix,ButchR_joinNMF-method}
\title{H-Matrix (H-Matrix with smallest frobError)}
\usage{
HMatrix(x, k = NULL, ...)

\S4method{HMatrix}{ButchR_NMF}(x, k = NULL, ...)

\S4method{HMatrix}{ButchR_integrativeNMF}(x, k = NULL, view_id = NULL, type = "all", ...)

\S4method{HMatrix}{ButchR_joinNMF}(x, k = NULL, ...)
}
\arguments{
\item{x}{an object of type ButchR_NMF, ButchR_joinNMF, or
ButchR_integrativeNMF.}

\item{k}{numeric - factorization rank.}

\item{...}{additional parameters.}

\item{view_id}{for integrative NMF; character vector with views
from which to extract H matrices}

\item{type}{for integrative NMF; type of H matrix to extract, could be:
\itemize{
\item shared - shared H matrix
\item viewspec - view specific H matrix
\item total - sum of shared H matrix and view specific H matrix.
\item all - shared H matrix and view specific H matrices.
}}
}
\value{
list of H-Matrices or an H-Matrix for the indicaded rank.
}
\description{
Return a list of H-Matrices or an H-Matrix for the indicaded rank
}
\examples{
data("leukemia")
nmf_exp <- run_NMF_tensor(leukemia$matrix, ranks = 2,
method = "NMF",
n_initializations = 2)
HMatrix(nmf_exp)
HMatrix(nmf_exp, k = 2)
HMatrix(nmf_exp)
HMatrix(nmf_exp, k = 2)
\dontrun{
# For ButchR_integrativeNMF objects:
# extract H matrices for all factorization ranks
HMatrix(inmf_exp, type = "shared")
HMatrix(inmf_exp, type = "viewspec")
HMatrix(inmf_exp, type = "total")
HMatrix(inmf_exp, type = "all")
# extract H matrices only for selected rank
HMatrix(inmf_exp, k = 2, type = "shared")
HMatrix(inmf_exp, k = 2, type = "viewspec")
HMatrix(inmf_exp, k = 2, type = "total")
HMatrix(inmf_exp, k = 2, type = "all")
# extract H matrices only for selected view and rank
HMatrix(inmf_exp, k = 2, view_id = "atac", type = "viewspec")
HMatrix(inmf_exp, k = 2, view_id = "atac", type = "total")
}

\dontrun{
# For ButchR_joinNMF objects:
HMatrix(jnmf_exp)
HMatrix(jnmf_exp, k = 2)
lapply(HMatrix(jnmf_exp, k = 2), head)
HMatrix(jnmf_exp, k = 2, view_id = "atac")
}
}
