% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrativeNMF-class.R
\docType{class}
\name{integrative_NMF-class}
\alias{integrative_NMF-class}
\alias{integrative_NMF}
\title{Integrative NMF Class}
\value{
An object of integrative_NMF Experiment Class
}
\description{
Integrative NMF Class
}
\section{Slots}{

\describe{
\item{\code{input_data}}{lsit of matrices}

\item{\code{HMatrix}}{list.}

\item{\code{HMatrix_vs}}{list.}

\item{\code{WMatrix_vs}}{list.}

\item{\code{FrobError}}{DataFrame.}

\item{\code{OptKStats}}{DataFrame.}

\item{\code{OptK}}{numeric}

\item{\code{SignFeatures}}{DataFrame or list}
}}

\examples{
\dontrun{
integrative_NMF(input_data   = input_data,
HMatrix      = shared_HMatrix_list,
HMatrix_vs   = view_specific_HMatrix_list,
}
}
