% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrativeNMF-class.R
\docType{class}
\name{integrative_NMF-class}
\alias{integrative_NMF-class}
\alias{integrative_NMF}
\title{Integrative NMF Class}
\value{

}
\description{
Integrative NMF Class
}
\section{Slots}{

\describe{
\item{\code{HMatrix}}{list.}

\item{\code{HMatrix_vs}}{list.}

\item{\code{WMatrix_vs}}{list.}

\item{\code{FrobError}}{DataFrame.}

\item{\code{OptKStats}}{DataFrame.}
}}

