% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinNMF.R
\docType{class}
\name{join_NMF-class}
\alias{join_NMF-class}
\alias{join_NMF}
\title{Join NMF Class}
\value{

}
\description{
Join NMF Class
}
\section{Slots}{

\describe{
\item{\code{shared_HMatrix_list}}{list.}

\item{\code{view_specific_WMatrix_list}}{list.}

\item{\code{best_factorization_idx}}{numeric}

\item{\code{view_specific_NMFexp_list}}{list}
}}

