% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinNMF.R
\docType{class}
\name{join_NMF-class}
\alias{join_NMF-class}
\alias{join_NMF}
\title{Join NMF Class}
\value{

}
\description{
Join NMF Class
}
\section{Slots}{

\describe{
\item{\code{HMatrix}}{list.}

\item{\code{WMatrix_vs}}{list.}

\item{\code{FrobError}}{DataFrame.}

\item{\code{OptKStats}}{DataFrame.}
}}

