% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/nmfExperiment-class_lite.R,
%   R/nmfExperiment-class.R
\docType{methods}
\name{normalizeH}
\alias{normalizeH}
\alias{normalizeH,nmfExperiment_lite-method}
\alias{normalizeH,ANY,ANY-method}
\alias{normalizeH,nmfExperiment-method}
\title{Normalize the signatures matrix (H)}
\usage{
normalizeH(nmf_exp, ...)

\S4method{normalizeH}{nmfExperiment_lite}(nmf_exp)

\S4method{normalizeH}{nmfExperiment}(nmf_exp)
}
\arguments{
\item{nmf_exp}{an nmfExperiment or a nmfExperiment_lite object}
}
\value{
an nmfExperiment or a nmfExperiment_lite object normalized by W
}
\description{
After row normalization of the matrix H, the inverse factors are
mutiplied with the columns of W in order to keep the matrix product W*H
constant.
}
\examples{
normalizeH(nmf_exp)
}
