% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/nmfExperiment-class.R,
%   R/nmfExperiment-class_lite.R
\docType{methods}
\name{normalizeW}
\alias{normalizeW}
\alias{normalizeW,nmfExperiment-method}
\alias{normalizeW,ANY,ANY-method}
\alias{normalizeW,nmfExperiment_lite-method}
\title{Normalize the signatures matrix (W)}
\usage{
normalizeW(nmf_exp, ...)

\S4method{normalizeW}{nmfExperiment}(nmf_exp)

\S4method{normalizeW}{nmfExperiment_lite}(nmf_exp)
}
\arguments{
\item{nmf_exp}{an nmfExperiment or a nmfExperiment_lite object}
}
\value{
an nmfExperiment or a nmfExperiment_lite object normalized by W
}
\description{
After column normalization of the matrix W, the inverse factors are
mutiplied with the rows of H in order to keep the matrix product W*H
constant.
}
\examples{
normalizeW(nmf_exp)
}
