% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfExperiment-class_lite.R
\docType{methods}
\name{OptK,nmfExperiment_lite-method}
\alias{OptK,nmfExperiment_lite-method}
\title{Return optimal K}
\usage{
\S4method{OptK}{nmfExperiment_lite}(x, ...)
}
\arguments{
\item{x}{an nmfExperiment_lite object}
}
\value{
numeric - optimal K
}
\description{
Return optimal K
}
\examples{
OptK(nmf_exp)
}
