% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/nmfExperiment-class_lite.R,
%   R/nmfExperiment-class.R
\docType{methods}
\name{regularizeH}
\alias{regularizeH}
\alias{regularizeH,nmfExperiment_lite-method}
\alias{regularizeH,ANY,ANY-method}
\alias{regularizeH,nmfExperiment-method}
\title{Regularize the signatures matrix (H)}
\usage{
regularizeH(nmf_exp, ...)

\S4method{regularizeH}{nmfExperiment_lite}(nmf_exp)

\S4method{regularizeH}{nmfExperiment}(nmf_exp)
}
\arguments{
\item{nmf_exp}{}
}
\value{
an nmfExperiment or a nmfExperiment_lite object normalized by W
}
\description{
After row regularization of the matrix H, the inverse factors are
mutiplied with the columns of W in order to keep the matrix product W*H
constant.
}
\examples{
regularizeH(nmf_exp)
}
