% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{WMatrix}
\alias{WMatrix}
\title{W-Matrix (W-Matrix with smallest frobError)}
\usage{
WMatrix(x, k = NULL, ...)
}
\arguments{
\item{x}{an nmfExperiment or a nmfExperiment_lite object}

\item{k}{numeric  - factorization rank}
}
\value{
list of W-Matrices or a W-Matrix for the indicaded rank
}
\description{
Return a list of W-Matrices or a W-Matrix for the indicaded rank
}
\examples{
WMatrix(nmf_exp)
WMatrix(nmf_exp, k = 2)
}
