% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/nmfExperiment-class_lite.R,
%   R/integrativeNMF-class.R, R/joinNMF-class.R
\docType{methods}
\name{WMatrix}
\alias{WMatrix}
\alias{WMatrix,nmfExperiment_lite-method}
\alias{WMatrix,ANY-method}
\alias{WMatrix,integrative_NMF-method}
\alias{WMatrix,join_NMF-method}
\title{W-Matrix (W-Matrix with smallest frobError)}
\usage{
WMatrix(x, k = NULL, ...)

\S4method{WMatrix}{nmfExperiment_lite}(x, k = NULL, ...)

\S4method{WMatrix}{integrative_NMF}(x, k = NULL, view_id = NULL, ...)

\S4method{WMatrix}{join_NMF}(x, k = NULL, view_id = NULL, ...)
}
\arguments{
\item{x}{an R object of type nmfExperiment, nmfExperiment_lite, join_NMF or
integrative_NMF.}

\item{k}{numeric - factorization rank.}

\item{view_id}{character vector with views from which to extract W matrices}
}
\value{
list of W-Matrices or a W-Matrix for the indicaded rank.
}
\description{
Return a list of W-Matrices or a W-Matrix for the indicaded rank
}
\examples{
WMatrix(nmf_exp)
WMatrix(nmf_exp, k = 2)
# For nmfExperiment_lite objects:
WMatrix(nmf_exp)
WMatrix(nmf_exp, k = 2)
# For integrative_NMF objects:
WMatrix(inmf_exp)
WMatrix(inmf_exp, k = 2)
lapply(WMatrix(inmf_exp, k = 2), head)
WMatrix(inmf_exp, k = 2, view_id = "atac")
# For join_NMF objects:
WMatrix(jnmf_exp)
WMatrix(jnmf_exp, k = 2)
lapply(WMatrix(jnmf_exp, k = 2), head)
WMatrix(jnmf_exp, k = 2, view_id = "atac")

}
