% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfExperiment-class_lite.R, R/tmp.R
\docType{methods}
\name{WMatrix,nmfExperiment_lite-method}
\alias{WMatrix,nmfExperiment_lite-method}
\title{Return a list of W-Matrices or a W-Matrix for the indicaded rank}
\usage{
\S4method{WMatrix}{nmfExperiment_lite}(x, k = NULL, ...)

\S4method{WMatrix}{nmfExperiment_lite}(x, k = NULL, ...)
}
\arguments{
\item{x}{an nmfExperiment_lite object}

\item{k}{numeric  - factorization rank}
}
\value{
list of W-Matrices or a W-Matrix for the indicaded rank
}
\description{
#----------------------------------------------------------------------------#
  #                                    Setup                                   #
  #----------------------------------------------------------------------------#
  # Check  data
  if (min(X) < 0 ) {
    stop("\nNegative values present in input matrix\n
         only non-negative matrices supported\n")
  }
  # Convert params to integer
  nmf_params <- lapply(list(ranks                 = ranks,
                            n_initializations     = n_initializations,
                            iterations            = iterations,
                            convergence_threshold = convergence_threshold),
                       as.integer)
}
\examples{
WMatrix(nmf_exp)
WMatrix(nmf_exp, k = 2)
}
